move transit trip leg builder guts to private header
